fix(state): normalize problem type hash in observation
 - all the values in the observation were nicely normalized to 0.0-1.0 except for the problem type hash which was a hash value and its negative, usually large values.
 - in tests normalizing the problem type caused agents to start grasping at good solutions around 250 episodes rather than more than 1000 episodes

BREAKING CHANGE: observation size is 2 floating point values more than before. In order to normalize problem type, we need more than two options or the normalized value will be likely to be 0.0 and 1.0 in some order. By adding two more variations on the input, we get a range of 4 values with at least 2 usually being not 1.0 or 0.0